split code into modules
closes #73
