+++ src/element/mod.rs
@@ -1,0 +1,4 @@
+pub mod traits;
+pub mod photo;
+pub mod collection;
+pub mod album;
